pyproject.toml
[tool.poetry]
name = "exam-reconciler"
version = "0.1.0"
description = "Exam topic reconciler agent"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "0.100.0"
uvicorn = {extras = ["standard"], version = "0.22.0"}
pandas = "2.2.2"
openpyxl = "3.1.2"
python-dotenv = "1.0.0"
supabase = "1.7.3"
psycopg = {extras = ["binary"], version = "3.2.0"}
sentence-transformers = "2.2.2"
pdfplumber = "0.9.0"
requests = "2.31.0"
aiolimiter = "1.0.0"

[tool.poetry.dev-dependencies]
pytest = "7.4.2"
ruff = "0.0.287"
black = "24.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
